% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_crps_news.R
\name{analyse_crps_news}
\alias{analyse_crps_news}
\title{Analyses a dataframe with news body for occurences of the each of the currencySymbols}
\usage{
analyse_crps_news(databodynews, currencySymbols)
}
\arguments{
\item{databodynews}{dataframe with news body and their timestamp}

\item{currencySymbols}{a vector of characters representing the cryptocurrencys Symbols to be analysed}
}
\value{
Dataframe with a boolean column per cryptosymbol, representing if the cryptocurrency was cited in the news article
}
\description{
Analyses a dataframe with news body for occurences of the each of the currencySymbols
}
