% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candle_plot.R
\name{candle_plot}
\alias{candle_plot}
\title{Candle plot}
\usage{
candle_plot(data, supp_type)
}
\arguments{
\item{data}{dataframe with the information}

\item{supp_type}{name of the column that has the supplementary information (volume or MACD)}
}
\value{
plotly object
}
\description{
Function to make an interactive plot
}
\details{
This function will return a plot about the prices of a certain coin along the time as well as some statistics such as the mean, the moving average; the volume or MACD
}
