% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\docType{data}
\name{coins}
\alias{coins}
\title{This is data to be included in my package}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 1 columns.}
\usage{
coins
}
\description{
This is data to be included in my package
}
\keyword{data}
