% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crp_dic.R
\docType{data}
\name{crp_dic}
\alias{crp_dic}
\title{Dictionary with names of each crptocurrency and its symbol}
\format{Dataframe}
\usage{
crp_dic
}
\description{
Dictionary with names of each crptocurrency and its symbol
}
\keyword{datasets}
