% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.R
\name{crypto}
\alias{crypto}
\title{Crypto}
\usage{
crypto(timeframe, firstDay, lastDay, cryptocurrency = "BTC",
  comparison = "USD", n_MA, n_quick_MACD, n_slow_MACD, n_signal_MACD)
}
\arguments{
\item{timeframe}{timeframe in hour or day}

\item{firstDay}{first day to analyse in dd/mm/yyyy format}

\item{lastDay}{last day to analyse in dd/mm/yyyy format}

\item{cryptocurrency}{cryptocurrency to analyse}

\item{comparison}{currency to be compared}

\item{n_MA}{Window Moving Average}

\item{n_quick_MACD}{quick MACD}

\item{n_slow_MACD}{slow MACD}

\item{n_signal_MACD}{signal MACD}
}
\value{
dataframe with the time, highest price, lowest price, open price, close price and financial indicators in the chosen timeframe
}
\description{
Master function
}
\details{
Master function which aggragate all the function of our package
}
\examples{
\dontrun{
crypto("hour", "01/08/2018", "01/10/2018", "BTC", "USD",5 , 26, 12, 9)
}
}
