% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_correlation.R
\name{crypto_correlation}
\alias{crypto_correlation}
\title{crypto_correlation}
\usage{
crypto_correlation(firstDay, lastDay, cryptoA, cryptoB)
}
\arguments{
\item{firstDay}{first day to analyse in dd/mm/yyyy format}

\item{lastDay}{last day to analyse in dd/mm/yyyy format}

\item{cryptoA}{first cryptocurrency to correlate}

\item{cryptoB}{second cryptocurrency to correlate}
}
\value{
correlation between crypto A and crypto B
}
\description{
Function to calculate the correlation between two crypto currencies
}
\details{
This function is designed to calcualte the correlation between two cryptocurrencies in a given timeframe
}
\examples{
\dontrun{
crypto_correlation("01/09/2018", "01/10/2018", "BTC", "ETH")
}
}
