% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CryptoCurrencyName.R
\name{CryptoCurrencyName}
\alias{CryptoCurrencyName}
\title{Returns the name of a cryptocurrency using its symbol. Uses the dataframe crp_dic.csv}
\usage{
CryptoCurrencyName(CryptoSymbol)
}
\arguments{
\item{CryptoSymbol}{a character representing the cryptocurrency Symbol}
}
\value{
Character representing the name of the wanted cryptocurrency
}
\description{
Returns the name of a cryptocurrency using its symbol. Uses the dataframe crp_dic.csv
}
