% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CryptonewsData.R
\docType{data}
\name{CryptonewsData}
\alias{CryptonewsData}
\title{Dataset with the articles and their time of release from 2013 to now}
\format{Dataframe}
\usage{
data(CryptonewsData)
}
\description{
Dataset with the articles and their time of release from 2013 to now
}
\keyword{datasets}
