% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CryptoNewsOccurencesDays.R
\docType{data}
\name{CryptoNewsOccurencesDays}
\alias{CryptoNewsOccurencesDays}
\title{Dataset with the number of occurences of cryptocurrencies in the news per day}
\format{Dataframe}
\usage{
data(CryptoNewsOccurencesDays)
}
\description{
Dataset with the number of occurences of cryptocurrencies in the news per day
}
\keyword{datasets}
