% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CryptoNewsOccurencesHour.R
\docType{data}
\name{CryptoNewsOccurencesHour}
\alias{CryptoNewsOccurencesHour}
\title{Dataset with the number of occurences of cryptocurrencies in the news per hour}
\format{Dataframe}
\usage{
data(CryptoNewsOccurencesHour)
}
\description{
Dataset with the number of occurences of cryptocurrencies in the news per hour
}
\keyword{datasets}
