% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_hour.R
\name{day_hour}
\alias{day_hour}
\title{day_hour}
\usage{
day_hour(timeframe, firstDay, lastDay, cryptocurrency = "BTC",
  comparison = "USD")
}
\arguments{
\item{timeframe}{timeframe in hour or day}

\item{firstDay}{first day to analyse in dd/mm/yyyy format : if the input is the same as the lastDay, will be set as the day before lastDay}

\item{lastDay}{last day to analyse in dd/mm/yyyy format}

\item{cryptocurrency}{cryptocurrency to analyse}

\item{comparison}{currency to be compared}
}
\value{
dataframe with the time, highest price, lowest price, open price, close price of the chosen timeframe

dataframe with all the information required
}
\description{
Function to get the information from the cryptocompare API
}
\details{
This function has been designed to get a dataframe from the cryptocurrency API with the highest, lowest, open and close price from the cryptocurrency chosen in hour or day timeframe.
}
\examples{
\dontrun{
exampleSameCryptocurrency <- day_hour("day", "01/12/2017", "01/08/2018", "BTC", "BTC")
}
}
