% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_data_from.R
\name{dl_data_from}
\alias{dl_data_from}
\title{Downloads the news about cryptocurrencies from CryptoCompare API}
\usage{
dl_data_from(timestp)
}
\arguments{
\item{timestp}{a character of numerical value representing time}
}
\value{
Dataframe with the time and body of the news released after timestamp
}
\description{
Downloads the news about cryptocurrencies from CryptoCompare API
}
