% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imp_Crp.R
\name{get_imp_Crp}
\alias{get_imp_Crp}
\title{Downloads the 20 cryptocurrencies that have the highest market capitalization from cryptocompare API}
\usage{
get_imp_Crp()
}
\value{
vector of characters representing the symbols of the cryptocurencies
}
\description{
Downloads the 20 cryptocurrencies that have the highest market capitalization from cryptocompare API
}
