% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastPriceBitfinex.R
\name{getLastPriceBitfinex}
\alias{getLastPriceBitfinex}
\title{getLastPriceBitfinex}
\usage{
getLastPriceBitfinex(cryptocurrency, comparison)
}
\arguments{
\item{cryptocurrency}{cryptocurrency id}

\item{comparison}{to be compared with}
}
\value{
string with the last price of Bitfinex, the criptocurrency and the comparison
}
\description{
Function to get the last price of the cryptocurrency in comparision with the desired currency of the platform Bitfinex
}
