% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastPriceMultiplePlatforms.R
\name{getLastPriceMultiplePlatforms}
\alias{getLastPriceMultiplePlatforms}
\title{getLastPriceMultiplePlatforms}
\usage{
getLastPriceMultiplePlatforms(cryptocurrency)
}
\arguments{
\item{cryptocurrency}{cryptocurrency id}
}
\value{
dataframe with prices on six platforms
}
\description{
Function used to get the last price of the imputted crypto on the 6 biggest platforms
}
\examples{
\dontrun{
getLastPriceMultiplePlatforms("BTC")
}
}
