% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastweek_minute.R
\name{lastweek_minute}
\alias{lastweek_minute}
\title{lastweek_minute}
\usage{
lastweek_minute(cryptocurrency = "BTC", comparison = "USD",
  only_two_minutes = FALSE)
}
\arguments{
\item{cryptocurrency}{cryptocurrency to analyse}

\item{comparison}{currency to be compared}

\item{only_two_minutes}{whether only output last two minutes data}
}
\value{
dataframe with the time, highest price, lowest price, open price, close price of the chosen timeframe

dataframe with all the information required
}
\description{
Function to get the information from the cryptocompare API per minute
}
\details{
This function has been designed to get a dataframe from the cryptocompare API with the highest, lowest, open and close price from the cryptocurrency chosen in minute of the last week.
}
\examples{
\dontrun{
lastweek_minute("BTC", "USD")
}
}
