% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastweek_news_counter.R
\name{lastweek_news_counter}
\alias{lastweek_news_counter}
\title{lastweek_news_counter}
\usage{
lastweek_news_counter(cryptocurrency = "BTC")
}
\arguments{
\item{cryptocurrency}{cryptocurrency}
}
\value{
dataframe with time and mentioned
}
\description{
Function to count the inputted crpytocurrency being mentioned hourly in last week
}
\details{
This function is designed for giving a delayed indicator for the given crypto in the most recent week
}
\examples{
\dontrun{
countNewsLastWeekBitcoin <- lastweek_news_counter("BTC")
}
}
