% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetUSDValue.R
\name{NetUSDValue}
\alias{NetUSDValue}
\title{NetUSDValue}
\usage{
NetUSDValue(pocket, day = "11/12/2018")
}
\arguments{
\item{pocket}{a vector contains the units of currencies you have}

\item{day}{the net value in this day}
}
\value{
a number indicates the net value
}
\description{
this function is used to calculate the net USD value in the pocket which is a vector contains different currencies you currently have
}
