% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lastweek.R
\name{plot_lastweek}
\alias{plot_lastweek}
\title{Plot trend and citations for last week}
\usage{
plot_lastweek(cryptocurrency = "BTC", comparison = "USD",
  grouping = "day")
}
\arguments{
\item{cryptocurrency}{the currency to examined}

\item{comparison}{the currency over which, the exchange rate will be shown}

\item{grouping}{the citations can be grouped in terms of "day", "12 hours", "6 hours", "3 hours", "hour".}
}
\value{
2 plots. One of the exchange rate and one with the number of citations.
}
\description{
A plot exhibitng the trend of the exchange rate between the currencies chosen is displayed.
Supplementary, the number of citations referring to the chosen currency from selected news sources is shown.
}
\examples{
\dontrun{

Call with default values:
plot_lastweek()

Specifying the currencies for the display:
plot_lastweek(cryptocurrency= "ETH", comparison= "EUR")

Specifying the grouping of the citations:
plot_lastweek(grouping= "6 hours")

}

}
\seealso{
\code{\link{candle_plot}}
}
