% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTspToHour.R
\name{resTspToHour}
\alias{resTspToHour}
\title{Analyses a dataframe of news article timestamp and occurences of currencySymbols and sums the number of occurences of each currency per hour}
\usage{
resTspToHour(resTsp, CrpSymbols)
}
\arguments{
\item{resTsp}{dataframe of news article timestamp and boolean occurences of currencySymbols}

\item{CrpSymbols}{a vector of characters representing the cryptocurrencys Symbols to be analysed}
}
\value{
Dataframe with a counter column per cryptosymbol, representing the number of times each cryptocurrency was cited in the news articles diring the hour
}
\description{
Analyses a dataframe of news article timestamp and occurences of currencySymbols and sums the number of occurences of each currency per hour
}
