% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumlastDay.R
\name{SumlastDay}
\alias{SumlastDay}
\title{Analyses a dataframe of news article timestamp and occurences of currencySymbols and sums the number of occurences of each currency during the day specified in argument}
\usage{
SumlastDay(dayTstp, CrpSymbols, BoolDataNews)
}
\arguments{
\item{dayTstp}{the timestamp of the hour to be analysed}

\item{CrpSymbols}{a vector of characters representing the cryptocurrencys Symbols to be analysed}

\item{BoolDataNews}{dataframe of news article timestamp and boolean occurences of currencySymbols}
}
\value{
Dataframe with a counter column per cryptosymbol, representing the number of times each cryptocurrency was cited in the news articles during the hour
}
\description{
Analyses a dataframe of news article timestamp and occurences of currencySymbols and sums the number of occurences of each currency during the day specified in argument
}
