% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumlastHour.R
\name{SumlastHour}
\alias{SumlastHour}
\title{Analyses a dataframe of news article timestamp and occurences of currencySymbols and sums the number of occurences of each currency during the hour specified in argument}
\usage{
SumlastHour(hourTstp, CrpSymbols, BoolDataNews)
}
\arguments{
\item{hourTstp}{the timestamp of the hour to be analysed}

\item{CrpSymbols}{a vector of characters representing the cryptocurrencys Symbols to be analysed}

\item{BoolDataNews}{dataframe of news article timestamp and boolean occurences of currencySymbols}
}
\value{
Dataframe with a counter column per cryptosymbol, representing the number of times each cryptocurrency was cited in the news articles during the hour
}
\description{
Analyses a dataframe of news article timestamp and occurences of currencySymbols and sums the number of occurences of each currency during the hour specified in argument
}
