% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{transaction}
\alias{transaction}
\title{transaction}
\usage{
transaction(pocket_log, unit = 1, buycurrency = "BTC",
  sellcurrency = "USD", day = "05/12/2018", allowNegative = FALSE)
}
\arguments{
\item{pocket_log}{a data.frame which have the columns date and different currencies}

\item{unit}{how many unit of buycurrency you want to buy}

\item{buycurrency}{the currency you want to buy}

\item{sellcurrency}{the currency you used to buy the buycurrency}

\item{day}{the day you do the transaction}

\item{allowNegative}{whether or not allow negative units happend}
}
\value{
a updated pocket_log
}
\description{
This function is used to make a transaction
}
\details{
This function takes the input pocket_log and record the transaction in the new row of the pocket_log
}
