% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tsp_data.R
\name{Tsp_data}
\alias{Tsp_data}
\title{Tsp_data}
\usage{
Tsp_data(timestp)
}
\arguments{
\item{timestp}{date}
}
\value{
the dataframe retrieve from the api
}
\description{
small function use to create a dataframe for the use of the function lastweek_news_counter
}
\details{
This function is used only to build the lastweek_news_counter function
}
