% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDayNewsData.R
\name{updateDayNewsData}
\alias{updateDayNewsData}
\title{Updates the dataset CryptoNewsAnalysedDays with the latest news from CryptoCompare API}
\usage{
updateDayNewsData()
}
\value{
updated dataframe
}
\description{
Updates the dataset CryptoNewsAnalysedDays with the latest news from CryptoCompare API
}
