% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_monthly_transformation.R
\name{weekly_monthly_transformation}
\alias{weekly_monthly_transformation}
\title{weekly_monthly_transformation}
\usage{
weekly_monthly_transformation(df, timeframe)
}
\arguments{
\item{df}{dataframe to transform}

\item{timeframe}{month or day}
}
\value{
dataframe with the time, highest price, lowest price, open price, close price in the chosen timeframe

dataframe transformed
}
\description{
Function to transform the timeframe of the dataset
}
\details{
This function has been designed to transform  the timeframe (monthly or weekly) of the dataset returned by the function day_hour.R
}
\examples{
\dontrun{
weekly_monthly_transformation(bitcoinVsDollarExampleDay, "week")
}
}
